refactor(FormHasil): Update type system ke decimal dengan bug highlight
Changes:
- Update: dataSkor Dictionary(Of String, Integer) → Double
- Update: maxSkorGrafik Integer → Double = 100.0
- Add: Decimal formatting ToString("F1") di grid dan grafik
- Improve: Konsistensi display skor di semua UI components

⚠️ KNOWN BUG - PERLU IMMEDIATE FIX:
  Highlighting juara (baris ~103-109) masih memiliki critical type mismatch:
  - CInt(teksSkor) mengkonversi "0.6" → 0 (salah!)
  - Perlu diubah: CInt → CDbl + epsilon comparison
  - Impact: Juara tidak pernah ter-highlight saat scoring decimal

Akan diperbaiki di commit berikutnya segera.

+++ sistem-pakar/ModuleInferensi.vb
@@ -32,7 +32,7 @@
             'jika user jawab "YA"
             If jawabanUser(i) = 1 Then
                 Dim idTopikSoal As String = dtPertanyaan.Rows(i)("Id_topik").ToString()
-                Dim bobotSoal As Double = Convert.ToInt32(dtPertanyaan.Rows(i)("bobot_pertanyaan"))
+                Dim bobotSoal As Double = Convert.ToDouble(dtPertanyaan.Rows(i)("bobot_pertanyaan"))
 
                 'tambahkan bobot ke topik yang sesuai
                 If skorTopik.ContainsKey(idTopikSoal) Then
@@ -45,7 +45,7 @@
 
         For Each item In skorTopik
             Dim id As String = item.Key
-            Dim skor As Integer = item.Value
+            Dim skor As Double = item.Value
 
             If namaTopik.ContainsKey(id) Then
                 Dim namaAsli As String = namaTopik(id)
